pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
django-crispy-forms = "^1.13.0"
django-admin-interface = "^0.18.4"
django-private-chat2 = "^1.0.2"
channels = "^3.0.4"
django-countries = "^7.2.1"
stripe = "^2.64.0"
stripe3 = "^0.0.0"
django-chat = "^0.1.0"
django-allauth = "^0.47.0"
python-decouple = "^3.5"
django-debug-toolbar = "^3.2.4"
django-zxcvbn-password = "^2.1.1"
django-markdownx = "^3.0.1"
django-preferences = "^1.0.0"
django-guest-user = "^0.5.2"
django-changelogs = "^0.1"
django-random-id-model = "^0.1.1"
ChatterBot = "^1.1.0"
[tool.poetry.dev-dependencies]
